pyproject.toml
[tool.poetry]
name = "smu-osp"
version = "0.1.0"
description = ""
authors = ["Youngmin Joo <ymjoo827@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-environ = "^0.11.2"
scrapy = "^2.12.0"
selenium = "^4.27.1"
bs4 = "^0.0.2"
scrapy-selenium = "^0.0.7"
scrapy-playwright = "^0.0.42"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-cors-headers = "^4.6.0"
pillow = "^11.1.0"
requests = "^2.32.3"
social-auth-app-django = "^5.4.2"
pymysql = "^1.1.1"
gunicorn = "^23.0.0"
playwright = "^1.50.0"
beautifulsoup4 = "^4.13.3"
celery = "^5.4.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
